+++ .pytest_tmp/test_detects_env_vars2/main.go
@@ -1,0 +1,6 @@
+package main
+import "os"
+func main() {
+    key := os.Getenv("API_KEY")
+    _ = key
+}